+++ visualization/app/codeCharta/ui/codeMap/codeMap.render.service.ts
@@ -74,7 +74,7 @@ export class CodeMapRenderService {
 		const appSettings = this.storeService.getState().appSettings
 		const showLabelNodeName = appSettings.showMetricLabelNodeName
 		const showLabelNodeMetric = appSettings.showMetricLabelNameValue
-		const hightestNode = this.getHighestNode(sortedNodes)
+		const highestNode = this.getHighestNode(sortedNodes)
 
 		this.codeMapLabelService.clearLabels()
 		if (showLabelNodeName || showLabelNodeMetric) {
@@ -89,7 +89,7 @@ export class CodeMapRenderService {
 							showNodeName: showLabelNodeName,
 							showNodeMetric: showLabelNodeMetric
 						},
-						hightestNode
+						highestNode
 					)
 					amountOfTopLabels -= 1
 				}
@@ -98,7 +98,7 @@ export class CodeMapRenderService {
 	}
 
 	private getHighestNode(sortedNodes: Node[]) {
-		const sortedNodeValues = sortedNodes.map(node => node.height)
+		const sortedNodeValues = sortedNodes.map(({ height, heightDelta }) => height + Math.abs(heightDelta ?? 0))
 		return Math.max(...sortedNodeValues)
 	}